+++ CHANGELOG.md
@@ -54,6 +54,9 @@ with the `YY.MINOR.MICRO` scheme.
 
 - Fixes unnecessary vshard router and storage config updates.
 
+- Fixed a bug which could sometimes lead to unconstrained stream of requests
+  from a failed instance (implemented expontential backoff for sentinel requests).
+
 ### Lua API
 
 - Remove `pico.exit` function as it is no longer used.